requirements.txt
dash-bootstrap-components==1.6.0
dash-leaflet==1.0.15
dash==2.18.2
db-dtypes==1.3.1
earthengine-api==1.4.3
Flask==3.0.3
geopandas==1.0.1
google-cloud-bigquery-storage==2.27.0
google-cloud-bigquery==3.27.0
httplib2==0.22.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
watchdog==5.0.2
gunicorn==23.0.0




dash_ag_grid

